rename getjson into getjsonfromurl for better undertanding
