+++ pages/experience.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import NavigationBox from "../src/components/NavigationBox";
 import WebglBox from "../src/components/webgl/WebglBox";
-import getjsonfromurl from "../src/util/getjson";
+import getjsonfromurl from "../src/util/getjsonfromurl";
 import React, { useState } from "react";
 
 const Content = styled.div`
